Set Connection `port` field type to `Int`

+++ model/src/main/kotlin/org/katan/model/Connection.kt
@@ -4,6 +4,6 @@ interface Connection {
 
     val host: String
 
-    val port: Short
+    val port: Int
 
 }
